# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.8.0
  description: API PNIGVS — portail public, référentiel établissements, auth et application métier
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /app/admin/users:
    get:
      operationId: listAdminUsers
      tags: [app]
      summary: Liste des comptes agents (administration DVS)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Accès réservé DVS
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createAdminUser
      tags: [app]
      summary: Créer un compte directeur / fondateur d'établissement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminUserBody"
      responses:
        "201":
          description: Compte créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "400":
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Accès réservé DVS
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /app/admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [app]
      summary: Activer ou désactiver un compte agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminUserBody"
      responses:
        "200":
          description: Compte mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "404":
          description: Compte introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Accès réservé DVS
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /app/dashboard:
    get:
      operationId: getAppDashboard
      tags: [app]
      summary: Tableau de bord métier
      description: |
        Retourne les indicateurs clés et le périmètre de l'utilisateur connecté.
        Les comptages sont filtrés selon le rôle et la portée géographique (DRENA, établissement).
      responses:
        "200":
          description: Tableau de bord
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppDashboardResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /app/establishments:
    get:
      operationId: listAppEstablishments
      tags: [app]
      summary: Annuaire établissements (périmètre utilisateur)
      description: |
        Liste paginée des établissements filtrée selon le rôle et la portée géographique
        de l'utilisateur connecté (établissement, DRENA, région ou national).
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/SearchParam"
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/DrenaParam"
        - $ref: "#/components/parameters/DdenaParam"
        - $ref: "#/components/parameters/DepartmentParam"
        - $ref: "#/components/parameters/LocalityParam"
        - $ref: "#/components/parameters/TypeParam"
        - $ref: "#/components/parameters/StatusParam"
        - $ref: "#/components/parameters/SortParam"
        - $ref: "#/components/parameters/OrderParam"
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstablishmentListResponse"
        "400":
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /app/establishments/{id}:
    get:
      operationId: getAppEstablishmentById
      tags: [app]
      summary: Fiche établissement (périmètre utilisateur)
      description: Retourne un établissement si l'utilisateur connecté y a accès selon son périmètre RBAC.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Établissement trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstablishmentDetail"
        "400":
          description: Identifiant invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Hors périmètre
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Établissement introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /activities:
    get:
      operationId: listActivities
      tags: [workflow]
      summary: Liste des activités scolaires
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createActivity
      tags: [workflow]
      summary: Créer une activité scolaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Activité créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivitySummary"
        "400":
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /activities/{id}:
    get:
      operationId: getActivityById
      tags: [workflow]
      summary: Détail d'une activité
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Activité trouvée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivitySummary"
        "404":
          description: Activité introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /requests:
    get:
      operationId: listRequests
      tags: [workflow]
      summary: Liste des demandes d'autorisation
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/RequestStatusParam"
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestListResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createRequest
      tags: [workflow]
      summary: Créer une demande (brouillon)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "201":
          description: Demande créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "400":
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /requests/{id}:
    get:
      operationId: getRequestById
      tags: [workflow]
      summary: Détail d'une demande avec historique
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Demande trouvée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "404":
          description: Demande introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /requests/{id}/transitions:
    post:
      operationId: transitionRequest
      tags: [workflow]
      summary: Transition workflow d'une demande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestTransitionBody"
      responses:
        "200":
          description: Transition appliquée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "400":
          description: Transition invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Action non autorisée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /media-publications/{id}/incident:
    post:
      operationId: reportMediaPublicationIncident
      tags: [media]
      summary: Signaler un incident survenu lors d'une sortie
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportMediaPublicationIncidentBody"
      responses:
        "200":
          description: Incident enregistré
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPublicationDetail"
        "400":
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Action non autorisée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Publication introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /media-publications:
    get:
      operationId: listMediaPublications
      tags: [media]
      summary: Liste des publications média (app)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/MediaPublicationStatusParam"
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPublicationListResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createMediaPublication
      tags: [media]
      summary: Créer un dossier média (brouillon)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaPublicationBody"
      responses:
        "201":
          description: Publication créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPublicationDetail"
        "400":
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Action non autorisée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /media-publications/{id}:
    get:
      operationId: getMediaPublicationById
      tags: [media]
      summary: Détail d'une publication média
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Publication trouvée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPublicationDetail"
        "404":
          description: Publication introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /media-publications/{id}/transitions:
    post:
      operationId: transitionMediaPublication
      tags: [media]
      summary: Transition workflow publication média
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaPublicationTransitionBody"
      responses:
        "200":
          description: Transition appliquée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPublicationDetail"
        "400":
          description: Transition invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Action non autorisée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [workflow]
      summary: Notifications de l'utilisateur connecté
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
        maxLength: 200
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required:
            - code
            - message
      required:
        - error
...
    AdminUserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserSummary"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - pagination
    CreateAdminUserBody:
      type: object
      properties:
        email:
          type: string
          format: email
        fullName:
          type: string
          minLength: 2
          maxLength: 255
        password:
          type: string
          minLength: 12
          maxLength: 128
        roleCode:
          $ref: "#/components/schemas/SchoolHeadRoleCode"
        establishmentId:
          type: string
          format: uuid
      required:
        - email
        - fullName
        - password
        - roleCode
        - establishmentId
...